+++ KaguyaProjectV2/KaguyaBot/Core/Commands/NSFW/Doujin.cs
@@ -13,7 +13,7 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.NSFW
 {
     public class Doujin : KaguyaBase
     {
-        [PremiumCommand]
+        [PremiumUserCommand]
         [NsfwCommand]
         [Command("Doujin")]
         [Summary("Displays a random doujin in chat.")]